Replace custom generators with ghga generators
